refactor: use vite in cli templates
Here switched to vite in all templates.
Followed migration instructions and templates.
Still no way to test before new versions are published to pnpm.
