+++ fixtures/webstudio-remix-netlify-functions/env.d.ts
@@ -1,2 +1,2 @@
-/// <reference types="@remix-run/dev" />
+/// <reference types="vite/client" />
 /// <reference types="@remix-run/node" />
