+++ fixtures/webstudio-remix-netlify-functions/netlify.toml
@@ -1,14 +1,16 @@
 [build]
-command = "remix build && cp _app_redirects public/_redirects"
-publish = "public"
+command = "npm run build"
+publish = "build/client"
 
 [dev]
 command = "npm run dev"
-targetPort = 3000
+framework = "vite"
+
+# Set immutable caching for static files, because they have fingerprinted filenames
 
 [[headers]]
 for = "/build/*"
 
 [headers.values]
-# Set to 60 seconds as an example. You can also add cache headers via Remix. See the documentation on [headers](https://remix.run/docs/en/v1/route/headers) in Remix.
-"Cache-Control" = "public, max-age=60, s-maxage=60"
+
+"Cache-Control" = "public, max-age=31560000, immutable"
