--- fixtures/webstudio-remix-netlify-functions/remix.config.js
@@ -1,22 +1,0 @@
-const baseConfig =
-  process.env.NODE_ENV === "production"
-    ? // when running the Netify CLI or building on Netlify, we want to use
-      {
-        server: "./server.js",
-        serverBuildPath: ".netlify/functions-internal/server.js",
-      }
-    : // otherwise support running remix dev, i.e. no custom server
-      undefined;
-
-/** @type {import('@remix-run/dev').AppConfig} */
-module.exports = {
-  ...baseConfig,
-  ignoredRouteFiles: ["**/.*"],
-  serverModuleFormat: "cjs",
-  serverDependenciesToBundle: [
-    /@webstudio-is\//,
-    "nanoid",
-    "change-case",
-    "title-case",
-  ],
-};
