--- fixtures/webstudio-remix-vercel/remix.config.js
@@ -1,11 +1,0 @@
-/** @type {import('@remix-run/dev').AppConfig} */
-module.exports = {
-  ignoredRouteFiles: ["**/.*"],
-  serverModuleFormat: "cjs",
-  serverDependenciesToBundle: [
-    /@webstudio-is\//,
-    "nanoid",
-    "change-case",
-    "title-case",
-  ],
-};
