+++ package.json
@@ -26,7 +26,6 @@
     "pre-commit": "./node_modules/.bin/nano-staged"
   },
   "devDependencies": {
-    "@storybook/react-vite": "^7.4.0",
     "@types/css-tree": "^2.3.1",
     "@webstudio-is/eslint-config-custom": "workspace:*",
     "esbuild": "^0.20.2",
@@ -37,7 +36,8 @@
     "storybook": "^7.4.0",
     "tsx": "^4.7.2",
     "turbo": "^1.8.6",
-    "typescript": "5.2.2"
+    "typescript": "5.2.2",
+    "vite": "^5.2.8"
   },
   "engines": {
     "node": "18",
