+++ packages/cli/templates/netlify-edge-functions/netlify.toml
@@ -1,7 +1,16 @@
 [build]
 command = "npm run build"
-publish = "public"
+publish = "build/client"
 
 [dev]
 command = "npm run dev"
-targetPort = 3000
+framework = "vite"
+
+# Set immutable caching for static files, because they have fingerprinted filenames
+
+[[headers]]
+for = "/build/*"
+
+[headers.values]
+
+"Cache-Control" = "public, max-age=31560000, immutable"
