+++ packages/cli/templates/netlify-edge-functions/package.json
@@ -1,6 +1,5 @@
 {
   "scripts": {
-    "predev": "node -e \"fs.rmSync('./public/_redirects', { recursive: true, force: true })\"",
     "start": "netlify serve"
   },
   "dependencies": {
