--- packages/cli/templates/netlify-functions/server.js
@@ -1,7 +1,0 @@
-import { createRequestHandler } from "@netlify/remix-adapter";
-import * as build from "@remix-run/dev/server-build";
-
-export const handler = createRequestHandler({
-  build,
-  mode: process.env.NODE_ENV,
-});
