+++ packages/cli/templates/netlify-functions/vite.config.ts
@@ -1,0 +1,16 @@
+import { resolve } from "node:path";
+import { vitePlugin as remix } from "@remix-run/dev";
+import { defineConfig } from "vite";
+import { netlifyPlugin } from "@netlify/remix-adapter/plugin";
+
+export default defineConfig({
+  plugins: [remix(), netlifyPlugin()],
+  resolve: {
+    alias: [
+      {
+        find: "~",
+        replacement: resolve("app"),
+      },
+    ],
+  },
+});
