+++ packages/cli/templates/vercel/package.json
@@ -1,5 +1,1 @@
-{
-  "scripts": {
-    "start": "remix-serve build"
-  }
-}
+{}
